# _backend/openapi.yaml
openapi: 3.0.0
info:
  title: CatTok
  description: >
    Made for **CatTok** frontend client, GitHub [littlefuny/cattok](https://github.com/littlefunky/cattok)

    <iframe src="https://giphy.com/embed/JuFwy0zPzd6jC" width="480" height="380" frameBorder="0" class="giphy-embed" allowFullScreen></iframe><p><a href="https://giphy.com/gifs/computer-computers-caturday-JuFwy0zPzd6jC">via GIPHY</a></p>
      
    This project was made for educational purpose only 🚀.
  version: 1.0.0
servers:
  - url: /v1
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
  parameters:
    user_id:
      name: user_id
      description: User ID or use alias `me` for authenticated user
      required: true
      in: path
      schema:
        type: string
        default: me
  responses:
    UnexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnauthorizedError:
      description: Unauthorized error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Fail"
          example:
            status: fail
            data: "Unauthorized"
  schemas:
...
    Success:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          default: success
        data:
          description: Returned data
          nullable: true
    Fail:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          default: fail
        data:
          description: Failed message
    Error:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          default: error
        message:
          description: Error message
          type: string
    User:
      type: object
      required:
        - id
        - name
        - email
        - photo
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        photo:
          type: string
    Post:
      type: object
      required:
        - id
        - user_id
        - title
        - description
        - video_url
        - public_date
        - heart
      properties:
        id:
          type: string
        user_id:
          type: string
        title:
          type: string
        description:
          type: string
        video_url:
          type: string
        public_date:
          type: string
          format: date-time
        heart:
          type: integer
...
paths:
  /user:
    post:
      tags:
        - User
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - photo
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                photo:
                  type: string
                  format: binary
      description: Create new user
      responses:
        "201":
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        required:
                          - jwt
                        properties:
                          jwt:
                            type: string
        "400":
          description: Malformed or missing required data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /user/{user_id}:
    get:
      tags:
        - User
      description: Get specific `User` by ID
      parameters:
        - $ref: "#/components/parameters/user_id"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        default:
          $ref: "#/components/responses/UnexpectedError"
    put:
      tags:
        - User
      description: Update user
      parameters:
        - $ref: "#/components/parameters/user_id"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                photo:
                  type: string
                  format: binary
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          description: Malformed or missing required data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "403":
          description: No permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /user/{user_id}/post:
    get:
      tags:
        - User
      description: Get `Post` associated with user
      parameters:
        - $ref: "#/components/parameters/user_id"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Post"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /login:
    post:
      tags:
        - User
      security: []
      description: Authenticate to CatTok
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          jwt:
                            type: string
        "400":
          description: Malformed or missing required data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        default:
          $ref: "#/components/responses/UnexpectedError"
...